addedEnsure old profile picture is deleted and fix default preview

+++ canvas_project/account_management/views.py
@@ -109,12 +109,14 @@ def update_account(request):
 
             profile, created = UserProfile.objects.get_or_create(user=user)
             if request.POST.get("delete_picture") == "1":
-                if profile.profile_picture:
-                    profile.profile_picture.delete()  # Datei aus MEDIA_ROOT löschen
-                profile.profile_picture = "profile_pics/default.jpg"  # Standardbild setzen
-
+                if profile.profile_picture and profile.profile_picture.name != "profile_pics/default.jpg":
+                    profile.profile_picture.delete()  # delete former profile picture
+                profile.profile_picture = "profile_pics/default.jpg"  # set default profile picture
             # Set profile picture only if a new one is uploaded
             elif form.cleaned_data.get("profile_picture"):
+                 # Check if the current profile picture exists and is not the default picture.
+                if profile.profile_picture and profile.profile_picture.name != "profile_pics/default.jpg":
+                    profile.profile_picture.delete()
                 profile.profile_picture = form.cleaned_data["profile_picture"]
 
             user.save()
